Componentize all parts of the review section
